`<InputArea>` - Enable e2e. Add test/utils/protractor. (#1639)
