+++ src/test-common.js
@@ -2,26 +2,17 @@ import omit from 'lodash/omit';
 
 // TODO: move to separate folders
 export {createDriverFactory} from 'wix-ui-test-utils/driver-factory';
+
+// TODO: Protractor helper should be imported from './test/utils/protractor'
 export {
-    getStoryUrl,
+    getStoryUrl, // Deprecated
     waitForVisibilityOf,
     protractorTestkitFactoryCreator,
     isFocused,
-    scrollToElement
-} from 'wix-ui-test-utils/protractor';
-
-// TODO: protractor helper methods should be moved to wix-ui-test-utils/protractor
-/**
- * Protractor helper for checking the presence of an attribute.
- */
-export const hasAttribute = (elementFinder, attributeName) =>
-  elementFinder.getAttribute(attributeName).then(value => value !== null);
-/**
- * Protractor helper for checking the presence of a css class name.
- */
-export const hasClass = (element, className) =>
-  element.getAttribute('class').then(classes => classes.split(' ').some(c => c.includes(className)));
-
+    scrollToElement,
+    hasAttribute,
+    hasClass
+} from '../test/utils/protractor';
 
 export {enzymeTestkitFactoryCreator} from 'wix-ui-test-utils/enzyme';
 export {testkitFactoryCreator} from 'wix-ui-test-utils/vanilla';
